Fix AI multicam exploits & a few other bugs
There's no longer a weird horrible frame
Space now renders (at the cost of some kinda weird lookin stuff in the multicam room but w/e)
Fixed an exploit involving getting AI eyes qdel'd that disabled static for all other eyes - this would allow you to see into maintenance, for instance, completely bypassing all AI limitations. HUGE exploit.

+++ code/modules/overmap/spacetravel.dm
@@ -60,7 +60,11 @@ proc/overmap_spacetravel(var/turf/space/T, var/atom/movable/A)
 	var/obj/effect/overmap/visitable/M = get_overmap_sector(T.z)
 	if (!M)
 		return
-
+	
+	// Don't let AI eyes yeet themselves off the map
+	if(istype(A, /mob/observer/eye))
+		return
+	
 	if(A.lost_in_space())
 		if(!QDELETED(A))
 			qdel(A)
